allow per-entry messages to be displayed to the user
You may want to have a message specifically regarding the branch
or commit that is blacklisted displayed to the user which is trying
to push it.  A comment on the line with the entry will now be printed
along with the template and error.

It's also possible to format this text through an external command
using the blacklist.formatcmd config setting.  Using this, you can
modify the way this text is displayed, such as with a silly cow:

git config blacklist.formatcmd "cowsay"
 _____________________________
< This branch is not allowed! >
 -----------------------------
        \   ^__^
         \  (oo)\_______
            (__)\       )\/\
                ||----w |
                ||     ||